+++ packages/flux-react/CHANGELOG.md
@@ -1,3 +1,8 @@
 # @psionic/flux-react
 
 ## 0.0.1
+
+<i>Oct 19, 2022</i>
+
+* 🥳 First publish of `@psionic/flux-react`.
+* ⚛️ Introduction of `useFluxReader` hook for interacting with Flux objects.